# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPay API",
        "description": "Documentation de l'API OMPay (Orange Money Pay)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        },
        {
            "url": "https://ompay.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/comptes/{id}/depot": {
            "post": {
                "tags": [
                    "Compte"
                ],
                "summary": "Effectuer un dépôt sur un compte",
                "operationId": "depotCompte",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "montant": {
                                        "type": "number",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué"
                    }
                }
            }
        },
        "/paiement-marchand": {
            "post": {
                "tags": [
                    "Compte"
                ],
                "summary": "Effectuer un paiement vers un marchand (par code_marchand ou téléphone)",
                "operationId": "paiementMarchand",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "montant": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "code_marchand": {
                                        "type": "string",
                                        "example": "MRC1234"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771428150"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement marchand effectué"
                    }
                }
            }
        },
        "/comptes/{id}/transfert": {
            "post": {
                "tags": [
                    "Compte"
                ],
                "summary": "Effectuer un transfert entre comptes",
                "operationId": "transfertCompte",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte source",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "montant": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "+221770112233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué"
                    }
                }
            }
        },
        "/comptes/{id}/solde": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Afficher le solde d'un compte",
                "operationId": "soldeCompte",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré"
                    }
                }
            }
        },
        "/comptes/{id}/transactions": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Afficher les transactions d'un compte",
                "operationId": "transactionsCompte",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "total_pages": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total_items": {
                                                            "type": "integer",
                                                            "example": 25
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Historique récupéré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comptes/{id}/profil": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Afficher le profil d'un compte",
                "operationId": "profilCompte",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil récupéré"
                    }
                }
            }
        },
...
    },
...
}